requirements.txt
# Core RAG System Dependencies
PyMuPDF==1.23.14
pytesseract==0.3.10
Pillow==10.1.0
numpy==1.24.3
pandas==2.1.4
sentence-transformers==2.2.2
chromadb==0.4.18
streamlit==1.28.1
tqdm==4.66.1
requests==2.31.0

# Additional dependencies
torch==2.1.1
transformers==4.36.2
tokenizers==0.15.0

# For OCR support (requires Tesseract installation)
# On Windows: Download from https://github.com/UB-Mannheim/tesseract/wiki
# On macOS: brew install tesseract
# On Ubuntu: sudo apt install tesseract-ocr

# Optional: GPU support for faster embeddings
# torch-audio==2.1.1
# torch-vision==0.16.1
